Refactor to standalone Angular components
Migrated app structure to use standalone Angular components. Renamed component files to follow the .component.* convention, updated imports, and replaced the root app.html/app.ts with app.component.html/app.component.ts. Updated dependencies in package.json and package-lock.json, including @angular/router and typescript.
