+++ cv-angular-2/src/app/app.component.html
@@ -1,0 +1,11 @@
+<app-header></app-header>
+<div class="container">
+  <app-aside></app-aside>
+  <main>
+    <app-main-content></app-main-content>
+    <app-tools-skills></app-tools-skills>
+    <app-education-experience></app-education-experience>
+  </main>
+</div>
+<app-footer></app-footer>
+
