+++ cv-angular-2/src/app/app.spec.ts
@@ -1,5 +1,5 @@
 import { TestBed } from '@angular/core/testing';
-import { App } from './app';
+import { App } from './app.component';
 
 describe('App', () => {
   beforeEach(async () => {
