+++ cv-angular-2/src/app/aside/aside.component.html
