+++ cv-angular-2/src/app/education-experience/education-experience.component.ts
@@ -3,8 +3,8 @@ import { Component } from '@angular/core';
 @Component({
   selector: 'app-education-experience',
   imports: [],
-  templateUrl: './education-experience.html',
-  styleUrl: './education-experience.css'
+  templateUrl: './education-experience.component.html',
+  styleUrl: './education-experience.component.css'
 })
 export class EducationExperience {
