+++ cv-angular-2/src/app/footer/footer.component.css
