+++ cv-angular-2/src/app/footer/footer.component.ts
@@ -3,8 +3,8 @@ import { Component } from '@angular/core';
 @Component({
   selector: 'app-footer',
   imports: [],
-  templateUrl: './footer.html',
-  styleUrl: './footer.css'
+  templateUrl: './footer.component.html',
+  styleUrl: './footer.component.css'
 })
 export class Footer {
