+++ cv-angular-2/src/app/main-content/main-content.component.css
