+++ cv-angular-2/src/app/main-content/main-content.component.ts
@@ -4,8 +4,8 @@ import { CommonModule } from '@angular/common';
 @Component({
   selector: 'app-main-content',
   imports: [CommonModule],
-  templateUrl: './main-content.html',
-  styleUrl: './main-content.css'
+  templateUrl: './main-content.component.html',
+  styleUrl: './main-content.component.css'
 })
 export class MainContent {
   carouselIndex = 0;
