+++ cv-angular-2/src/app/tools-skills/tools-skills.component.css
