+++ cv-angular-2/src/app/tools-skills/tools-skills.component.ts
@@ -3,8 +3,8 @@ import { Component } from '@angular/core';
 @Component({
   selector: 'app-tools-skills',
   imports: [],
-  templateUrl: './tools-skills.html',
-  styleUrl: './tools-skills.css'
+  templateUrl: './tools-skills.component.html',
+  styleUrl: './tools-skills.component.css'
 })
 export class ToolsSkills {
