+++ cv-angular-2/src/index.html
@@ -9,13 +9,5 @@
 </head>
 <body>
   <app-root></app-root>
-  <app-header></app-header>
-<main>
-  <app-main-content></app-main-content>
-  <app-aside></app-aside>
-</main>
-<app-tools-skills></app-tools-skills>
-<app-education-experience></app-education-experience>
-<app-footer></app-footer>
 </body>
 </html>
